backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.1.0
python-decouple==3.8
psycopg2-binary==2.9.9
gunicorn==21.2.0
requests==2.31.0
numpy==1.24.3
pandas==2.0.3
redis==5.0.1
celery==5.3.4
